Assert that transmitCommandOrPoll()'s destination is redundant
The destination parameter appears to be redundant.  Add an assert to
highlight any cases where it isn't set as expected.

If this assertion never fails, it would be useful to remove the
parameter to simplify callers and make it clearer that there is a
one-to-one mapping between each physical terminal and its
corresponding instance of this class.

Removing the parameter may also be useful for any future attempts to
separate the interface to the converter hardware and the VT52
emulation into separate processes, where it would be preferable for
any given process to not be able to send commands to a terminal
managed by another process.