Add custom terminfo file and 122KEY_EN_CUSTOM scancode mapping
Add a terminfo file which enables the use of the F1 through F12 keys
which were recently added to the 122KEY_EN scancode mapping, and
supports the Backspace key generating a delete character.

Add variant of the 122KEY_EN scancode mapping which generates a delete
character when Backspace is pressed.

Add a script which compiles the terminfo file.

A subsequent commit will add an optional script that uses the terminfo
file from this location.