Accept a login/shell path via -l parameter; add default and example scripts
Update the -l parameter to require a path to a login or shell
executable.  Add a new etc/twinax_login_default script which is used
as the default value for that parameter.  Users passing `-l` will need
to instead pass `-l /usr/bin/login` (or whatever path is used by their
distribution).

Set an additional TWINAXTERM environment variable before running the
login or shell executable.

Provide alternative scripts which use the custom terminfo file and
optionally the default or a custom .bashrc.