122KEY_EN: Fix blank keys at the top of the numeric keypad
Fix the scan codes expected for the top row of (blank) keys on the
numeric keypad to avoid overwriting the entry for the ` key.
Previously the ` key generated a "*" character in both unshifted and
shifted states due to the erroneous entry, and none of the keys in the
top row of the numeric keypad generated characters.