122KEY_EN: Add/fix generation of some control characters in the range 0x00..0x1F
Generate character 0x00 (^@) when the space bar is pressed with Ctrl
held, which is consistent with xterm and rxvt-unicode.

Generate character 0x1F (^_) when the - key is pressed with Ctrl held.
xterm only generates that character if Shift is also held, but the
combination of Ctrl and Shift isn't supported by the scancode mapping
table.  rxvt-unicode doesn't require Shift to be held to generate that
character.  The same character was already generated when the / key is
pressed with Ctrl held, which is also consistent with both xterm and
rxvt-unicode.

To enable 0x1F to be generated by the above key combination, move
generation of character 0x1C (^\) to the \ key when Ctrl is held,
which is consistent with xterm and rxvt-unicode.

Generating character 0x1C from the above key combination means that it
can no longer generate character 0x1D (^]).  The latter will still be
generated by the ¢ key when Ctrl is held; that key generates ] when
Shift is held, and Ctrl + ] generates 0x1D in xterm and rxvt-unicode,
so this is somewhat consistent.

Generate character 0x1E (^^) when 6 is pressed with Ctrl held.  This
is consistent with xterm and rxvt-unicode.

It appears that various other Ctrl + key combinations generate 0x1C
through 0x1F in xterm, so this is still not an exhaustive emulation of
standard terminal behavior.