122KEY_EN: Generate < and > from shifted , and . keys
Generate "<" and ">" from the shifted state of the , and . keys for
consistency with a standard US English 101-key keyboard.  While those
characters are also available via the < key on the US English 122-key
keyboard, use of that key requires learning new behavior since it's
not present on a 101-key keyboard, whereas the ";" and ":" which were
previously generated in the shifted state of these keys are available
via the nearby ; key on both the 101-key and 122-key keyboards.